+++ offchain/eth-input-reader/src/config.rs
@@ -20,12 +20,10 @@ use types::deployment_files::{
     rollups_deployment::{RollupsDeployment, RollupsDeploymentJson},
 };
 
-use crate::auth::{AuthConfig, AuthEnvCLIConfig, AuthError};
-
-#[derive(Parser)]
+#[derive(Clone, Parser)]
 #[command(name = "rd_config")]
-#[command(about = "Configuration for dispatcher")]
-pub struct DispatcherEnvCLIConfig {
+#[command(about = "Configuration for rollups eth-input-reader")]
+pub struct EthInputReaderEnvCLIConfig {
     #[command(flatten)]
     pub sc_config: SCEnvCLIConfig,
 
@@ -35,12 +33,6 @@ pub struct DispatcherEnvCLIConfig {
     #[command(flatten)]
     pub broker_config: BrokerCLIConfig,
 
-    #[command(flatten)]
-    pub auth_config: AuthEnvCLIConfig,
-
-    #[command(flatten)]
-    pub log_config: LogEnvCliConfig,
-
     /// Path to file with deployment json of dapp
     #[arg(long, env, default_value = "./dapp_deployment.json")]
     pub rd_dapp_deployment_file: PathBuf,
@@ -49,18 +41,16 @@ pub struct DispatcherEnvCLIConfig {
     #[arg(long, env, default_value = "./rollups_deployment.json")]
     pub rd_rollups_deployment_file: PathBuf,
 
-    /// Duration of rollups epoch in seconds, for which dispatcher will make claims.
+    /// Duration of rollups epoch in seconds, for which eth-input-reader will make claims.
     #[arg(long, env, default_value = "604800")]
     pub rd_epoch_duration: u64,
 }
 
 #[derive(Clone, Debug)]
-pub struct DispatcherConfig {
+pub struct EthInputReaderConfig {
     pub sc_config: SCConfig,
     pub tx_config: TxManagerConfig,
     pub broker_config: BrokerConfig,
-    pub auth_config: AuthConfig,
-    pub log_config: LogConfig,
 
     pub dapp_deployment: DappDeployment,
     pub rollups_deployment: RollupsDeployment,
@@ -82,9 +72,6 @@ pub enum Error {
         source: std::io::Error,
     },
 
-    #[snafu(display("Auth configuration error: {}", source))]
-    AuthError { source: AuthError },
-
     #[snafu(display("Json parse error ({})", path.display()))]
     JsonParseError {
         path: PathBuf,
@@ -100,56 +87,51 @@ pub enum Error {
 
 #[derive(Debug)]
 pub struct Config {
-    pub dispatcher_config: DispatcherConfig,
+    pub eth_input_reader_config: EthInputReaderConfig,
     pub http_server_config: HttpServerConfig,
 }
 
 impl Config {
     pub fn initialize() -> Result<Self, Error> {
-        let (http_server_config, dispatcher_config) =
-            HttpServerConfig::parse::<DispatcherEnvCLIConfig>("dispatcher");
+        let (http_server_config, eth_input_reader_config) =
+            HttpServerConfig::parse::<EthInputReaderEnvCLIConfig>(
+                "eth_input_reader",
+            );
 
-        let sc_config = SCConfig::initialize(dispatcher_config.sc_config)
+        let sc_config = SCConfig::initialize(eth_input_reader_config.sc_config)
             .context(StateClientSnafu)?;
 
         let tx_config =
-            TxManagerConfig::initialize(dispatcher_config.tx_config)
+            TxManagerConfig::initialize(eth_input_reader_config.tx_config)
                 .context(TxManagerSnafu)?;
 
-        let auth_config = AuthConfig::initialize(dispatcher_config.auth_config)
-            .context(AuthSnafu)?;
-
-        let log_config = LogConfig::initialize(dispatcher_config.log_config);
-
-        let path = dispatcher_config.rd_dapp_deployment_file;
+        let path = eth_input_reader_config.rd_dapp_deployment_file;
         let dapp_deployment: DappDeployment = read_json(path)?;
 
-        let path = dispatcher_config.rd_rollups_deployment_file;
+        let path = eth_input_reader_config.rd_rollups_deployment_file;
         let rollups_deployment = read_json::<RollupsDeploymentJson>(path)
             .map(RollupsDeployment::from)?;
 
-        let broker_config = BrokerConfig::from(dispatcher_config.broker_config);
+        let broker_config =
+            BrokerConfig::from(eth_input_reader_config.broker_config);
 
         assert!(
             sc_config.default_confirmations < tx_config.default_confirmations,
             "`state-client confirmations` has to be less than `tx-manager confirmations,`"
         );
 
-        let dispatcher_config = DispatcherConfig {
+        let eth_input_reader_config = EthInputReaderConfig {
             sc_config,
             tx_config,
             broker_config,
-            auth_config,
-            log_config,
-
             dapp_deployment,
             rollups_deployment,
-            epoch_duration: dispatcher_config.rd_epoch_duration,
+            epoch_duration: eth_input_reader_config.rd_epoch_duration,
             priority: Priority::Normal,
         };
 
         Ok(Config {
-            dispatcher_config,
+            eth_input_reader_config,
             http_server_config,
         })
     }