+++ offchain/eth-input-reader/src/eth_input_reader.rs
@@ -9,12 +9,13 @@ use tracing::{error, info, instrument, trace, warn};
 use types::foldables::authority::rollups::{RollupsInitialState, RollupsState};
 
 use crate::{
-    config::DispatcherConfig,
-    drivers::{blockchain::BlockchainDriver, machine::MachineDriver, Context},
-    error::{BrokerSnafu, DispatcherError, SenderSnafu, StateServerSnafu},
-    machine::{rollups_broker::BrokerFacade, BrokerReceive, BrokerSend},
-    metrics::DispatcherMetrics,
-    sender::ClaimSender,
+    config::EthInputReaderConfig,
+    error::{BrokerSnafu, EthInputReaderError, StateServerSnafu},
+    machine::{
+        driver::MachineDriver, rollups_broker::BrokerFacade, BrokerReceive,
+        BrokerSend, Context,
+    },
+    metrics::EthInputReaderMetrics,
     setup::{create_block_subscription, create_context, create_state_server},
 };
 
@@ -22,21 +23,15 @@ use snafu::{whatever, ResultExt};
 
 #[instrument(level = "trace", skip_all)]
 pub async fn start(
-    config: DispatcherConfig,
-    metrics: DispatcherMetrics,
-) -> Result<(), DispatcherError> {
-    info!("Setting up dispatcher");
+    config: EthInputReaderConfig,
+    metrics: EthInputReaderMetrics,
+) -> Result<(), EthInputReaderError> {
+    info!("Setting up eth-input-reader with config: {:?}", config);
 
     let dapp_metadata = DAppMetadata {
         chain_id: config.tx_config.chain_id,
         dapp_address: Address::new(config.dapp_deployment.dapp_address.into()),
     };
-
-    trace!("Creating transaction manager");
-    let mut claim_sender =
-        ClaimSender::new(&config, dapp_metadata.clone(), metrics.clone())
-            .await
-            .context(SenderSnafu)?;
 
     trace!("Creating state-server connection");
     let state_server = create_state_server(&config.sc_config).await?;
@@ -62,15 +57,13 @@ pub async fn start(
     trace!("Creating machine driver and blockchain driver");
     let mut machine_driver =
         MachineDriver::new(config.dapp_deployment.dapp_address);
-    let mut blockchain_driver =
-        BlockchainDriver::new(config.dapp_deployment.dapp_address);
 
     let initial_state = RollupsInitialState {
         history_address: config.rollups_deployment.history_address,
         input_box_address: config.rollups_deployment.input_box_address,
     };
 
-    trace!("Starting dispatcher...");
+    trace!("Starting eth-input-reader...");
     loop {
         match block_subscription.next().await {
             Some(Ok(BlockStreamItem::NewBlock(b))) => {
@@ -81,15 +74,13 @@ pub async fn start(
                     b.hash,
                     b.parent_hash
                 );
-                claim_sender = process_block(
+                process_block(
                     &b,
                     &state_server,
                     &initial_state,
                     &mut context,
                     &mut machine_driver,
-                    &mut blockchain_driver,
                     &broker,
-                    claim_sender,
                 )
                 .await?
             }
@@ -133,12 +124,9 @@ async fn process_block(
 
     context: &mut Context,
     machine_driver: &mut MachineDriver,
-    blockchain_driver: &mut BlockchainDriver,
 
     broker: &(impl BrokerSend + BrokerReceive),
-
-    claim_sender: ClaimSender,
-) -> Result<ClaimSender, DispatcherError> {
+) -> Result<(), EthInputReaderError> {
     trace!("Querying rollup state");
     let state = state_server
         .query_state(initial_state, block.hash)
@@ -152,9 +140,5 @@ async fn process_block(
         .await
         .context(BrokerSnafu)?;
 
-    // Drive blockchain
-    trace!("Reacting to state with `blockchain_driver`");
-    blockchain_driver
-        .react(&state.state.history, broker, claim_sender)
-        .await
+    Ok(())
 }