+++ CHANGELOG.md
@@ -63,6 +63,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed timestamp in GraphQL API
 - Fixed BigInt in GraphQL API
 
+### Changed
+
+- Renamed `dispatcher` service to `eth-input-reader`
+
+### Removed
+
+- Removed claiming functionality from `eth-input-reader`
+
 ## [1.0.0] 2023-08-22
 
 ### Added